Moved generic functions to APIs utils/bids
